requirements.txt
paho-mqtt==1.5.1
click==7.1.2
dnspython3
python-dateutil
pyYAML==5.3.1
protobuf==3.14.0
requests==2.25.1